Finished adding auth to all endpoints.

+++ src/tasks/tasks.service.ts
@@ -10,20 +10,14 @@ import { TaskStatus } from './tasks.status.enum';
 @Injectable()
 export class TasksService {
 
-    constructor(
-        @InjectRepository(TaskRepository)
-        private taskRepository: TaskRepository
-    ) {}
+    constructor( @InjectRepository(TaskRepository) private taskRepository: TaskRepository) {}
 
-    async getTasks(
-        filterDTO: GetTasksFilterDTO,
-        user: User
-    ): Promise<Task[]> {
+    async getTasks(filterDTO: GetTasksFilterDTO, user: User): Promise<Task[]> {
         return await this.taskRepository.getTasks(filterDTO, user);
     }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(id: number, user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id }});
 
         if (!found) {
             throw new NotFoundException(`Task with id "${id}" not found`);
@@ -32,23 +26,20 @@ export class TasksService {
         return found;
     }
 
-    async createTask(
-        createTaskDTO: CreateTaksDTO,
-        user: User,
-    ): Promise<Task> {
+    async createTask(createTaskDTO: CreateTaksDTO, user: User): Promise<Task> {
         return await this.taskRepository.createTask(createTaskDTO, user);
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({id, userId: user.id});
 
         if (result.affected === 0) {
             throw new NotFoundException(`Task with id "${id}" not found`);
         }
     }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+        const task = await this.getTaskById(id, user);
 
         task.status = status;
